requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
faiss-cpu
google-generativeai
